feature: finaliza controlador de notas fiscais

+++ src/Estoque/Estoque.Application/DTOs/ProdutoDTO.cs
@@ -4,5 +4,5 @@ public class ProdutoDTO
 {
     public string Codigo { get; set; }
     public string Descricao { get; set; }
-    public decimal Saldo { get; set; }
+    public int Saldo { get; set; }
 }
